+++ Library/src/Multiplayer/RealtimeEngineUtils.cpp
@@ -94,7 +94,8 @@ csp::multiplayer::SpaceEntity* FindSpaceObject(csp::common::IRealtimeEngine& Rea
 std::unique_ptr<csp::multiplayer::SpaceEntity> BuildNewAvatar(const csp::common::String& UserId, csp::common::IRealtimeEngine& RealtimeEngine,
     csp::common::IJSScriptRunner& ScriptRunner, csp::common::LogSystem& LogSystem, uint64_t NetworkId, const csp::common::String& Name,
     const csp::multiplayer::SpaceTransform& Transform, bool IsVisible, uint64_t OwnerId, bool IsTransferable, bool IsPersistent,
-    const csp::common::String& AvatarId, csp::multiplayer::AvatarState AvatarState, csp::multiplayer::AvatarPlayMode AvatarPlayMode)
+    const csp::common::String& AvatarId, csp::multiplayer::AvatarState AvatarState, csp::multiplayer::AvatarPlayMode AvatarPlayMode,
+    csp::multiplayer::LocomotionModel LocomotionModel)
 {
     auto NewAvatar = std::unique_ptr<csp::multiplayer::SpaceEntity>(new csp::multiplayer::SpaceEntity(
         &RealtimeEngine, ScriptRunner, &LogSystem, SpaceEntityType::Avatar, NetworkId, Name, Transform, OwnerId, {}, IsTransferable, IsPersistent));
@@ -103,6 +104,7 @@ std::unique_ptr<csp::multiplayer::SpaceEntity> BuildNewAvatar(const csp::common:
     AvatarComponent->SetAvatarId(AvatarId);
     AvatarComponent->SetState(AvatarState);
     AvatarComponent->SetAvatarPlayMode(AvatarPlayMode);
+    AvatarComponent->SetLocomotionModel(LocomotionModel);
     AvatarComponent->SetUserId(UserId);
     AvatarComponent->SetIsVisible(IsVisible);