+++ Tests/src/PublicAPITests/ComponentTests/PortalComponentTests.cpp
@@ -84,7 +84,7 @@ CSP_PUBLIC_TEST(CSPEngine, PortalTests, UsePortalTest)
         const auto LoginState = UserSystem->GetLoginState();
 
         auto [Avatar] = AWAIT(RealtimeEngine.get(), CreateAvatar, UserName, LoginState.UserId, UserTransform, IsVisible, UserAvatarState,
-            UserAvatarId, UserAvatarPlayMode);
+            UserAvatarId, UserAvatarPlayMode, LocomotionModel::Grounded);
 
         // Create object to represent the portal
         csp::common::String ObjectName = "Object 1";
@@ -117,7 +117,7 @@ CSP_PUBLIC_TEST(CSPEngine, PortalTests, UsePortalTest)
         const auto LoginState = UserSystem->GetLoginState();
 
         auto [Avatar] = AWAIT(RealtimeEngine.get(), CreateAvatar, UserName, LoginState.UserId, UserTransform, IsVisible, UserAvatarState,
-            UserAvatarId, UserAvatarPlayMode);
+            UserAvatarId, UserAvatarPlayMode, LocomotionModel::Grounded);
 
         auto [ExitSpaceResult] = AWAIT_PRE(SpaceSystem, ExitSpace, RequestPredicate);
     }