+++ .github/CHANGELOG.md
@@ -13,6 +13,11 @@ All notable changes to this project will be documented in this file.
 
 - [NT-0] refac: Moved enum StereoVideoType to SharedEnums StereoVideoType by MAG-JamesEdgeworth
   This change moves the StereoVideoType enum from the VideoComponent to the SharedEnums header, making it more accessible for use across different components and systems that may need to reference stereo video types.
+🔥 ❗Breaking Changes
+
+- [OF-1821] feat!: Pass `LocomotionModel` on init via `CreateAvatar` by mag-lt
+  When creating an Avatar, a `csp::multiplayer::LocomotionModel` must now be passed. This is a breaking change.
+  To obtain the same behaviour as before, `csp::multiplayer::LocomotionModel::Grounded` should be provided.
 
 ## [6.23.0] - 2026-02-04_17-54-18